BOAC-5189, prep StudentEnrollmentTerm component for fancy expand/collapse courses
