+++ src/components/student/profile/StudentEnrollmentTerm.vue
@@ -1,26 +1,12 @@
 <template>
   <v-card
-    :class="{'bg-light-blue student-term-current': config.currentEnrollmentTermId === parseInt(term.termId)}"
+    :class="{'bg-sky-blue student-term-current': config.currentEnrollmentTermId === parseInt(term.termId)}"
     density="compact"
     elevation="0"
     min-width="300"
   >
     <v-card-title class="pt-3 student-term-header">
-      <div class="align-baseline d-flex flex-wrap mb-1">
-        <h3 :id="`term-${term.termId}-header`" class="font-size-18">{{ term.termName }}</h3>
-        <div v-if="isConcurrent" class="font-size-14 text-grey-darken-2">&nbsp;UCBX</div>
-        <StudentAcademicStanding
-          v-if="term.academicStanding"
-          class="font-size-14 ml-1"
-          :standing="term.academicStanding"
-        />
-        <StudentWithdrawalCancel
-          v-if="student.sisProfile.withdrawalCancel"
-          class="font-size-14 ml-1"
-          :term-id="term.termId"
-          :withdrawal="student.sisProfile.withdrawalCancel"
-        />
-      </div>
+      <StudentEnrollmentTermAcademicStanding :student="student" :term="term" />
     </v-card-title>
     <v-card-text class="pb-3">
       <div role="table">
@@ -61,56 +47,17 @@
       </div>
     </v-card-text>
     <v-card-subtitle class="pb-3">
-      <div class="pt-3 student-term-footer">
-        <div class="d-flex justify-space-between">
-          <div :id="`term-${term.termId}-gpa`">
-            <span class="student-course-label">Term GPA: </span>
-            <span
-              v-if="round(get(term, 'termGpa.gpa', 0), 3) > 0"
-              class="font-size-14"
-            >
-              {{ round(get(term, 'termGpa.gpa', 0), 3) }}
-            </span>
-            <span v-else>&mdash;</span>
-          </div>
-          <div :id="`term-${term.termId}-units`" class="align-center d-flex justify-content-end">
-            <div class="student-course-label align-right">Total Units: </div>
-            <div class="font-size-14 text-right" :class="{'units-total': showMinUnits || showMaxUnits}">
-              <span v-if="get(term, 'enrolledUnits', 0) !== 0">{{ numFormat(term.enrolledUnits, '0.0') }}</span>
-              <span v-else>&mdash;</span>
-            </div>
-          </div>
-        </div>
-        <div
-          v-if="showMinUnits || showMaxUnits"
-          :id="`term-${term.termId}-units-allowed`"
-          class="text-right"
-        >
-          <div v-if="showMinUnits" class="align-center d-flex justify-content-end">
-            <div class="student-course-label align-right">Exception Min Units: </div>
-            <div :id="`term-${term.termId}-min-units`" class="font-size-14 units-total">
-              {{ numFormat(term.minTermUnitsAllowed, '0.0') || '&mdash;' }}
-            </div>
-          </div>
-          <div v-if="showMaxUnits" class="align-center d-flex justify-content-end">
-            <div class="student-course-label align-right">Exception Max Units: </div>
-            <div :id="`term-${term.termId}-max-units`" class="font-size-14 units-total">
-              {{ numFormat(term.maxTermUnitsAllowed, '0.0') || '&mdash;' }}
-            </div>
-          </div>
-        </div>
-      </div>
+      <StudentEnrollmentTermUnits :term="term" />
     </v-card-subtitle>
   </v-card>
 </template>
 
 <script setup>
-import StudentAcademicStanding from '@/components/student/profile/StudentAcademicStanding'
 import StudentCourse from '@/components/student/profile/StudentCourse'
-import StudentWithdrawalCancel from '@/components/student/profile/StudentWithdrawalCancel'
+import StudentEnrollmentTermAcademicStanding from '@/components/student/profile/StudentEnrollmentTermAcademicStanding.vue'
+import StudentEnrollmentTermUnits from '@/components/student/profile/StudentEnrollmentTermUnits.vue'
 import {DateTime} from 'luxon'
-import {get, isEmpty, isNil, some} from 'lodash'
-import {numFormat, round} from '@/lib/utils'
+import {isEmpty} from 'lodash'
 import {useContextStore} from '@/stores/context'
 
 const props = defineProps({
@@ -125,11 +72,8 @@ const props = defineProps({
 })
 const config = useContextStore().config
 const currentUser = useContextStore().currentUser
-const maxUnits = props.term.maxTermUnitsAllowed
-const minUnits = props.term.minTermUnitsAllowed
-const isConcurrent = some(props.term.enrollments, {'academicCareer': 'UCBX'})
-const showMaxUnits = isNil(maxUnits) && maxUnits !== config.defaultTermUnitsAllowed.max
-const showMinUnits = !isNil(minUnits) && minUnits !== config.defaultTermUnitsAllowed.min
+const student = props.student
+const term = props.term
 </script>
 
 <style scoped>
@@ -168,11 +112,5 @@ const showMinUnits = !isNil(minUnits) && minUnits !== config.defaultTermUnitsAll
   display: flex;
   flex-wrap: wrap;
   font-weight: 700;
-}
-.student-term-footer {
-  border-top: 1px #999 solid !important;
-}
-.units-total {
-  min-width: 30px;
 }
 </style>
